Add files via upload
After exploratory data analysis, I found that there are two growth parameters for trees
(circumference and height). These can be used for getting the life stages of a tree. As the focus
was on optimizing tree maintenance, descriptive statistical analysis was done. After removal
of outliers and binning of data according to circumference, visualizations were created to show
the regions with the number of trees classified in various growth stages.